+++ cypress/support/commands.js
@@ -1,30 +1,30 @@
-// ***********************************************
-// This example commands.js shows you how to
-// create various custom commands and overwrite
-// existing commands.
-//
-// For more comprehensive examples of custom
-// commands please read more here:
-// https://on.cypress.io/custom-commands
-// ***********************************************
-//
-//
+// // ***********************************************
+// // This example commands.js shows you how to
+// // create various custom commands and overwrite
+// // existing commands.
+
+// // For more comprehensive examples of custom
+// // commands please read more here:
+// // https://on.cypress.io/custom-commands
+// // ***********************************************
+
+
 // -- This is a parent command --
 // Cypress.Commands.add('login', (email, password) => { ... })
-//
-//
+
+
 // -- This is a child command --
 // Cypress.Commands.add('drag', { prevSubject: 'element'}, (subject, options) => { ... })
-//
-//
+
+
 // -- This is a dual command --
 // Cypress.Commands.add('dismiss', { prevSubject: 'optional'}, (subject, options) => { ... })
-//
-//
+
+
 // -- This will overwrite an existing command --
 // Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
-before(() => {
-  const { PORT = 3000 } = process.env;
-  cy.visit(`http://localhost:${PORT}`);
-  cy.waitForReact();
-});
+// before(() => {
+//   const { PORT = 3000 } = process.env;
+//   cy.visit(`http://localhost:${PORT}`);
+//   cy.waitForReact();
+// });
