+++ src/Components/Attendees.js
@@ -1,3 +1,5 @@
 export default function Attendees() {
-  return;
-}
+    return(
+        <></>
+    )
+  }
