+++ src/Components/Footer.js
@@ -1,3 +1,5 @@
 export default function Footer() {
-  return;
+  return(
+    <></>
+  )
 }
