+++ src/Components/Header.js
@@ -1,3 +1,5 @@
 export default function Header() {
-  return;
+  return(
+    <></>
+  )
 }
