+++ src/Components/NewEventForm.js
@@ -1,3 +1,23 @@
+import { MouseEvent } from "react";
+import 
+
+
 export default function NewEventForm({ handleAddEvent }) {
-  return;
+  return(
+    
+  
+  function addEvent() {
+    const createEvent = {
+      id: generateUniqueID(),
+      eventType: selectOption,
+      name: newEvent.name,
+      organizer: newEvent.organizer,
+      eventImage: newEvent.eventImage || "https://loremflickr.com/640/480/",
+      date: newEvent.date,
+      people: [],
+    };
+  </>
 }
+
+</>
+
